+++ src/lib/karma.ts
@@ -72,23 +72,16 @@ const defaultConfig: EnhancedConfigOptions = {
   },
   // see https://github.com/karma-runner/karma/issues/2119#issuecomment-239615791
   formatError(msg) {
-    return msg
-      .split('\n')
-      .reduce(
-        (list, line) => {
-          // filter node_modules
-          if (line.includes('/~/')) {
-            return list;
-          }
+    // filter out empty lines and node_modules
+    if (!msg.trim() || /~/.test(msg) || /node_modules/.test(msg)) return '';
 
-          // show only source line (without webpack protocol)
-          const sourceLine = line.split(' <- ')[0].replace('webpack:///', './');
-          list.push(`    ${sourceLine}`);
-          return list;
-        },
-        [] as Array<string>
-      )
-      .join('\n');
+    // change: "at Context.<anonymous> (webpack:///tests/unit.ts:6:26 <- dist-tests/index.js:1104:127)"
+    //     to: "at Context.<anonymous> (tests/unit.ts:6:26)"
+    const line = msg.includes('webpack:///')
+      ? msg.split(' <- ')[0].replace('webpack:///', '') + ')'
+      : msg;
+
+    return `  ${line}\n`;
   }
 };